fix(protocol): 64-bit offset math in sd_msg_recv_audio_body (32-bit OOB write)
The audio-reassembly bounds check computed the destination offset as
`(size_t)seq * SD_CHUNK_SIZE`. `seq` is peer-controlled and size_t is
only 32-bit on the ESP32, so any seq >= 2^20 overflows the multiply and
wraps the offset to a small in-range value — defeating the
`off + dlen > pcm_cap` check and turning an out-of-range sequence number
into an out-of-bounds memcpy into the PCM buffer on real hardware. The
host build never caught it: size_t is 64-bit there, so the offset never
wraps and the existing test (seq=1000) passes, giving false confidence.

Compute the offset in uint64_t (UINT32_MAX * 4096 + 4096 fits in 64
bits), so the bounds check is sound on both the 64-bit host and the
32-bit target. Behavior on the host is unchanged.

Adds test_audio_body_seq_offset_overflow: seq = 2^20 (seq*4096 == 2^32,
which wraps to 0 in 32-bit arithmetic) into a buffer larger than one
chunk, so a wrapped offset of 0 would slip past the size check — only
honest 64-bit math rejects it. Locks the contract against regression.

+++ firmware/components/protocol/src/messages.c
@@ -171,13 +171,20 @@ sd_err_t sd_msg_recv_audio_body(sd_transport_t *t, uint32_t chunks, uint32_t siz
         uint32_t seq = ((uint32_t)frame[0] << 24) | ((uint32_t)frame[1] << 16) |
                        ((uint32_t)frame[2] << 8) | (uint32_t)frame[3];
         size_t dlen = len - 4;
-        size_t off = (size_t)seq * SD_CHUNK_SIZE;
+        /* Compute the destination offset in 64-bit. `seq` is peer-controlled and
+         * size_t is only 32-bit on the ESP32, so `(size_t)seq * SD_CHUNK_SIZE`
+         * would wrap (seq >= 2^20 overflows 32 bits) and defeat the bounds check
+         * below — turning an out-of-range sequence number into an out-of-bounds
+         * memcpy. 64-bit math cannot overflow here (UINT32_MAX * 4096 + 4096 fits
+         * in uint64_t), so the check is sound on both the 64-bit host and the
+         * 32-bit target. */
+        uint64_t off = (uint64_t)seq * SD_CHUNK_SIZE;
         if (off + dlen > pcm_cap) {
             err = SD_ERR_IO; /* out-of-range sequence number */
             goto done;
         }
         if (dlen > 0) {
-            memcpy(pcm + off, frame + 4, dlen);
+            memcpy(pcm + (size_t)off, frame + 4, dlen);
         }
         total += dlen;
     }
